Fix race in TestInputSteerer
The assertion

    require.Equal(t, "b", string(p2[:n]))

would not always succeed, because the InputSteerer may have read "b"
from the input before the test pushed the new writer.

Fix this by allowing the reads from the input to progress one byte at a
time, and blocking the read of the second byte until the new writer has
been pushed.