remove async solves for remote-directories
The ast.Directory Close is never called, so that ends up holding
the Solves running for the entire build.  See:
https://github.com/openllb/hlb/pull/323#issuecomment-1190974238

For now removing Close from the interface until we can figure out how
to more narrowly scope the remote directory usage.  This also moves
the file lookups to be synchronous on use (so still lazy if never used).
This could have a large impact if we were to access many files from
the remote directory, but typically I think we use this to just
import the `module.hlb` file for a remote import.

Future optimizations might be to ref-count the solve so we can reuse
the single solve for the life time of all the files returned from
Directory.Open.