Set implicit local directory path to ""
Previously, this was being set to ".", and would wrongly strip the first
"." from a path that begins with "." in ModuleDir. Setting it to the
empty string seems more correct: nothing will be stripped. The only
other use of this "root" is in filepath.Join calls inside localDirectory
methods, which will also do the right thing with an empty string.

Fixes #332