Include last modified time in LocalUniqueID
I ran into a situation where a "local" mount was wrongly deduped with a
"local" from a concurrent solve that was created by another hlb process
before some files in that dir were modified.

By including the last modified timestamp in the unique ID, we an avoid
this deduplication in the case where a file has actually changed.