use ForceNoOutput for mounts that are not bound
We dont care about mutations to mounts unless they are bound with
`as name` syntax.  However, buildkit does not know this so it assumes
that mounts are mutable and will may be used elsewhere, so anytime we
use a mount (including `local`) it will prevent buildkit from caching
the vertex.